+++ test/memhouse/f2_transactional_writes_audit_jobs_test.exs
@@ -45,8 +45,7 @@ defmodule VanishingSubjectProvider do
              "confidence_percentage" => 90,
              "sensitivity" => "internal",
              "target_level" => "peer",
-             "update_operation" => "add",
-             "hearsay" => false
+             "update_operation" => "add"
            }
          ]
        },